Create README.md
# N322

[Final Project](https://in-info-web4.informatics.iupui.edu/~smccalle/WarDragonsCraftingTracker/www/)

## This project is an ionic app that allows players of the mobile game War Dragons to track their gear for the gear crafting event.
A user can look through the types of gear and see the basic stats abput the gear, such as shard requirements to craft and level, types of elements, the time needed to craft and level the gear, and the amount of rubies that can be used to speed up the crafting time. 

## The user, after logging in, can save an instance of gear to keep track of levels and to plan what to level in crafting by seeing the requirements for their desired level.
The app can keep track of the users shard inventory and compare to the requirements for leveling so the user knows if they have enough shards to level. This way the user can plan which pieces of gear to level for efficient leveling.

The user can have multiple sets of gear, so the app allows them to distinguish sets by naming the instance.

## Create
the user creates an instance of gear in which they can keep track of current levels for each piece of gear

## Read
the user can view each gear type and read their instances of sets of gear

## Update
the user can update their instance of gear, primarily through the level of the gear

## Delete
The user can delete an instance of gear they created